refactor: Constant 클래스의 패키지 변경
util 패키지로 이동
