+++ src/main/java/store/view/OutputView.java
@@ -1,12 +1,11 @@
 package store.view;
 
-import static store.view.Constant.BLANK;
-import static store.view.Constant.INTRODUCE_PRODUCT_MESSAGE;
+import static store.util.Constant.BLANK;
+import static store.util.Constant.INTRODUCE_PRODUCT_MESSAGE;
 import static store.util.ErrorMessage.ERROR_TAG;
-import static store.view.Constant.RECEIPT_AMOUNT_INFO_TITLE;
-import static store.view.Constant.RECEIPT_GIVEAWAY_TITLE;
-import static store.view.Constant.RECEIPT_PRODUCT_NAME_QUANTITY_AMOUNT;
-import static store.view.Constant.RECEIPT_TITLE;
+import static store.util.Constant.RECEIPT_AMOUNT_INFO_TITLE;
+import static store.util.Constant.RECEIPT_GIVEAWAY_TITLE;
+import static store.util.Constant.RECEIPT_TITLE;
 import static store.view.MessageTemplate.PRODUCT_INFO;
 import static store.view.MessageTemplate.PRODUCT_INFO_ZERO_QUANTITY;
 import static store.view.MessageTemplate.RECEIPT_GIVEAWAY_HISTORY;
